// frontend/src/types/bowlers.ts
export type bowlers = {
    bowlerID: number;
    bowlerLastName: string;
    bowlerMiddleInit: string;
    bowlerFirstName: string;
    bowlerAddress: string;
    bowlerCity: string;
    bowlerState: string;
    bowlerZip: number;
    bowlerPhoneNumber: string;
    teamName: string;
}

export type team = {
    teamid: number;
    teamname: string;
  };